Add test call for credit control request.
Add a client call to send a credit control request. Will eventually add a http endpoint that can be hit to trigger it. Fixed the response, if it doesn't have the correct fields, it will bail out and the client will get a timeout. Need to deal with the error, maybe in the handle_error in the credit control callback.